change something

+++ tests/am_cputests/src/cputests/mul_u64.rs
@@ -11,7 +11,7 @@ use alloc::{
 
 #[no_mangle]
 pub struct MulU64Test {
-    test_data: Vec<u64>,
+    test_data: Vec<u32>,
     answer: Vec<u64>,
 }
 
@@ -34,7 +34,7 @@ impl BenchMark for MulU64Test {
         for i in 0..self.test_data.len() {
             for j in i..self.test_data.len() {
                 println!("0x{:x} * 0x{:x} = 0x{:x}, answer: 0x{:x}", self.test_data[i], self.test_data[j], self.test_data[i] * self.test_data[j], self.answer[ans_index]);
-                xs_assert_eq!(self.test_data[i] * self.test_data[j], self.answer[ans_index], self.err_type());
+                xs_assert_eq!(self.test_data[i] as u64 * self.test_data[j] as u64, self.answer[ans_index], self.err_type());
                 ans_index += 1;
             }
         }
@@ -46,7 +46,7 @@ impl BenchMark for MulU64Test {
             let mut ans_index = 0;
             for i in 0..self.test_data.len() {
                 for j in i..self.test_data.len() {
-                    xs_assert_eq!(self.test_data[i] * self.test_data[j], self.answer[ans_index], self.err_type());
+                    xs_assert_eq!(self.test_data[i] as u64 * self.test_data[j] as u64, self.answer[ans_index], self.err_type());
                     ans_index += 1;
                 }
             }
